Added a basic test and began adding LLVM things.
