+++ .gitignore
@@ -1,2 +1,9 @@
+# JetBrains IDE project files
 .idea/*
+
+# Build directory
 build/
+
+# Generated bytecode LL files
+*.bc
+*.ll
